Enable deep sleep
The keyboard is preventing my mac from going to sleep for very long. Deep sleep is not configured
by default upstream yet. Sounds like the battery reporting might be the things that keep waking
the machine up. I've also seen references to &sleep in Discord. Want to see if I can find that at
some point it the future and put that under LY_CTL.